Fix handling of autorefs
This commit refactors the handling of bom-refs in the CDX serializer.
The previous versions assumed that a protobom being serialized into CDX
could have refless nodes.

This edge case is not possible, the whole protobom graph is keyed on the
node ids.

What can happen is that a node may have been assigned an autogenerated ID
from a cdx component without a ref. When such a node is found, the serializer
will now remove the bom-ref to maintain the highest fidelity with the original
document.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>