Restructure flatstrings and add Checksum method to Node
This commit expands the flat string used to all substructures. The previous
iteration was flaky and was leaving out data.

This string is used to compute a Node checksum, and this commit adds a new
public method node.Checksum() used to access the computed checksum externally.

The implementation is still hidden away.

Signed-off-by: Adolfo Garcia Veytia (puerco) <puerco@chainguard.dev>