Add Initial Conformance Skeleton
This commit adds the initial conformance test skeleton.

Ath this time it is really simple, it has two simple tests using
the built in Equal() operator of the NodeList and a comparison of nodes.

The idea of this checkin is to skaffold the way we run conformance tests,
find golden data, etc. This initial check of the conformance testing
framwework should let us start adding more conformance tests incrementally.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>