Alpha SPDX 3.0 Serializer
This commit adds the beginning of an SPDX 3.0 serializer.

It is not enabled by default, applications wishing to use it
need to purposedly include the package and it will auto-register
with the protobom writer.

This package is not intended to be accurate yet, the custom SPDX3
types are private and are intended to only cover the fields needed
by protobom to render its data structures. The data assignments are
not yet complete and we may need to define additional elements.

This serializer will produce a nested JSON-LD document. My plan is
to add an option that configures the serializer to output the SBOM in
framed JSON-LD mode.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>