CDX(u): Make autogenerated IDs deterministic
This commit modifies the cyclondx component unserializer functions
to generate a determinisitic ID when parsing the SBOMs. Previously
we were using the randomized IDs leading to problems testing to
ensure the unserializer and serializers are cnoformant.

We also now flag the autogenereated IDs as si to ensure they are
removed when rendering the document.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>