Formalize SPDX namespace in serializer
This commit modifies the SPDX namespace behavior when serializing
documents. This change properly handles the URL fragments when writing
the SPDX document, stripping it from the doc when defined.

It introduces two options into the serializer to ensure that we
can control when things go wrong: FailOnInvalidDocIdFragment and
GenerateDocumentId, documentation on the serializer option is
in the next commits.

Signed-off-by: Adolfo García Veytia (puerco) <puerco@stacklok.com>